+++ aiu/youtube.py
@@ -1,15 +1,12 @@
-import io
 import json
-import os
-import requests
 import tempfile
 from typing import Tuple, TYPE_CHECKING
 from urllib.parse import parse_qs, urlparse
 
-from PIL import Image
 from ytm import YouTubeMusic, YouTubeMusicDL  # noqa
 
 from aiu import LOGGER
+from aiu.parser import fetch_image
 from aiu.typedefs import Duration
 
 if TYPE_CHECKING:
@@ -28,8 +25,8 @@ def get_album_id(link):
     return album
 
 
-def update_metadata(meta):
-    # type: (JSON) -> Tuple[str, JSON]
+def update_metadata(meta, fetch_cover=False):
+    # type: (JSON, bool) -> Tuple[str, JSON]
     LOGGER.debug("Updating YouTube Music album metadata")
     album_cover = meta["thumbnail"]
     album_cover = album_cover.get("path", album_cover["url"])
@@ -37,11 +34,12 @@ def update_metadata(meta):
         song["title"] = song["name"]
         song["track"] = track
         song["duration"] = str(Duration(int(song["length"] / 1000.0)))
-        song["cover"] = album_cover
         song["date"] = meta["date"]
         song["year"] = meta["date"]["year"]
         song["album"] = meta["name"]
         song["artist"] = meta["artists"][0]["name"]
+        if fetch_cover or not album_cover.startswith("http"):
+            song["cover"] = album_cover
     with tempfile.NamedTemporaryFile("w") as file:
         json.dump(meta, file, indent=4, ensure_ascii=False)
     return file.name, meta["tracks"]
@@ -53,7 +51,7 @@ def get_metadata(link):
     LOGGER.debug("Retrieving metadata from link: [%s]", link)
     api = YouTubeMusic()
     meta = api.album(album)
-    return update_metadata(meta)
+    return update_metadata(meta, fetch_cover=False)
 
 
 def fetch_files(link, output_dir, with_cover=True):
@@ -64,10 +62,6 @@ def fetch_files(link, output_dir, with_cover=True):
     meta = api.download_album(album, output_dir)  # pre-applied ID3 tags
     if with_cover:
         url = meta["thumbnail"]["url"]
-        path = os.path.join(output_dir, "cover.png")
-        resp = requests.get(url)
-        buffer = io.BytesIO(resp.content)
-        image = Image.open(buffer)
-        image.save(path, format="PNG")
+        path = fetch_image(url)
         meta["thumbnail"]["path"] = path
     return update_metadata(meta)